Atualiza dependências e códigos
