+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## WHAT'S NEW
 
+### Version 1.0.3
+* Updated Flutter dependencies to latest versions
+* Modernized image editing and sharing utilities
+
 Thanks for updating to the latest Version 1.0.2 DocLense!
 As always, we are grateful for your continued interest and support.
