+++ lib/ui_components/image_view.dart
@@ -6,38 +6,30 @@ import 'package:doclense/configs/app_typography.dart';
 import 'package:doclense/configs/space.dart';
 import 'package:doclense/constants/appstrings.dart';
 import 'package:doclense/constants/route_constants.dart';
-import 'package:doclense/utils/share_util_new.dart';
+import 'package:doclense/providers/image_list.dart';
+import 'package:doclense/providers/theme_provider.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_spinkit/flutter_spinkit.dart';
-import 'package:image/image.dart' as image_lib;
 import 'package:image_cropper/image_cropper.dart';
-import 'package:path/path.dart';
 import 'package:image_editor_plus/image_editor_plus.dart' hide ImageCropper;
+import 'package:path/path.dart';
 import 'package:provider/provider.dart';
-
-import '../providers/image_list.dart';
-import '../providers/theme_provider.dart';
-
-enum IconOptions { share }
 
 class Imageview extends StatefulWidget {
   final File file;
   final ImageList list;
 
-  const Imageview(
-    this.file,
-    this.list,
-  );
+  const Imageview(this.file, this.list);
 
   @override
   _ImageviewState createState() => _ImageviewState();
 }
 
 class _ImageviewState extends State<Imageview> {
-  CroppedFile? cropped;
   bool _isLoading = true;
   List<File> files = [];
   int index = 0;
+
   @override
   void initState() {
     super.initState();
@@ -47,14 +39,13 @@ class _ImageviewState extends State<Imageview> {
         _isLoading = false;
       });
     });
-  }  
-  
+  }
+
   Future<void> cropimage(File file, Color appBarColor, Color bgColor) async {
     if (await file.exists()) {
       try {
         final croppedFile = await ImageCropper().cropImage(
           sourcePath: file.path,
-          cropStyle: CropStyle.rectangle,
           compressQuality: 80,
           uiSettings: [
             AndroidUiSettings(
@@ -65,7 +56,7 @@ class _ImageviewState extends State<Imageview> {
             ),
           ],
         );
-        
+
         setState(() {
           if (croppedFile != null) {
             files.add(File(croppedFile.path));
@@ -77,51 +68,9 @@ class _ImageviewState extends State<Imageview> {
         // Fallback ao editor de imagem se o recorte falhar
         await getFilterImage(context, appBarColor);
       }
-        }
-
-        if (cropped != null) {
-          index++;
-          files.add(File(cropped!.path));
-        } else {
-          // files.add(file);
-        }
-      });
     }
   }
 
-  // Widget popupMenuButton() {
-  //   return PopupMenuButton<IconOptions>(
-  //     icon: const Icon(Icons.more_vert),
-  //     itemBuilder: (BuildContext context) => <PopupMenuEntry<IconOptions>>[
-  //       PopupMenuItem<IconOptions>(
-  //         value: IconOptions.share,
-  //         child: Row(children: const [
-  //           Icon(
-  //             Icons.share,
-  //             size: 28.0,
-  //             // color: Colors.blue,
-  //           ),
-  //           SizedBox(
-  //             width: 23.0,
-  //           ),
-  //           Text(
-  //             'Share',
-  //             style: TextStyle(fontSize: 20.0),
-  //           )
-  //         ]),
-  //       )
-  //     ],
-  //     onSelected: (IconOptions value) {
-  //       setState(() {
-  //         if (value == IconOptions.share) {
-  //           if (files.isNotEmpty) {
-  //             ShareUtil.sharePdf(files[index].path, 'Compartilhar PDF');
-  //           }
-  //         }
-  //       });
-  //     },
-  //   );
-  // }
   Future<void> getFilterImage(BuildContext context, Color appBarColor) async {
     File filterfile;
     if (files.isNotEmpty) {
@@ -131,19 +80,18 @@ class _ImageviewState extends State<Imageview> {
     }
 
     // Usando o ImageEditor diretamente
-    final Uint8List? editedImage = await Navigator.push(
+    final editedImage = await Navigator.push(
       context,
       MaterialPageRoute(
-        builder: (context) => ImageEditor(
-          image: filterfile.readAsBytesSync(),
-          appBarColor: appBarColor,
-        ),
+        builder: (context) => ImageEditor(image: filterfile.readAsBytesSync()),
       ),
     );
 
     if (editedImage != null) {
       // Salvar a imagem editada
-      final editedFile = File(filterfile.path.replaceFirst('.jpg', '_edited.jpg'));
+      final editedFile = File(
+        filterfile.path.replaceFirst('.jpg', '_edited.jpg'),
+      );
       await editedFile.writeAsBytes(editedImage);
 
       setState(() {
@@ -156,47 +104,48 @@ class _ImageviewState extends State<Imageview> {
 
   Future<void> _showChoiceDialogHome(BuildContext context) {
     return showDialog(
-        context: context,
-        builder: (BuildContext ctx) {
-          return AlertDialog(
-            backgroundColor: Colors.blueGrey[800],
-            title: Text(
-              S.deleteWarning,
-              textAlign: TextAlign.center,
-              style: AppText.b1!.cl(Colors.white),
-            ),
-            content: SingleChildScrollView(
-              child: ListBody(
-                children: <Widget>[
-                  GestureDetector(
-                    onTap: () {
-                      widget.list.imagelist = [];
-                      widget.list.imagepath = [];
+      context: context,
+      builder: (BuildContext ctx) {
+        return AlertDialog(
+          backgroundColor: Colors.blueGrey[800],
+          title: Text(
+            S.deleteWarning,
+            textAlign: TextAlign.center,
+            style: AppText.b1!.cl(Colors.white),
+          ),
+          content: SingleChildScrollView(
+            child: ListBody(
+              children: <Widget>[
+                GestureDetector(
+                  onTap: () {
+                    widget.list.imagelist = [];
+                    widget.list.imagepath = [];
 
-                      Navigator.of(context).popUntil((route) => route.isFirst);
-                    },
-                    child: Text(
-                      S.yes,
-                      textAlign: TextAlign.center,
-                      style: AppText.b1!.cl(Colors.white),
-                    ),
+                    Navigator.of(context).popUntil((route) => route.isFirst);
+                  },
+                  child: Text(
+                    S.yes,
+                    textAlign: TextAlign.center,
+                    style: AppText.b1!.cl(Colors.white),
                   ),
-                  Space.y!,
-                  GestureDetector(
-                    onTap: () {
-                      Navigator.of(ctx).pop();
-                    },
-                    child: Text(
-                      S.no,
-                      textAlign: TextAlign.center,
-                      style: AppText.b1!.cl(Colors.white),
-                    ),
+                ),
+                Space.y!,
+                GestureDetector(
+                  onTap: () {
+                    Navigator.of(ctx).pop();
+                  },
+                  child: Text(
+                    S.no,
+                    textAlign: TextAlign.center,
+                    style: AppText.b1!.cl(Colors.white),
                   ),
-                ],
-              ),
+                ),
+              ],
             ),
-          );
-        });
+          ),
+        );
+      },
+    );
   }
 
   @override
@@ -208,79 +157,40 @@ class _ImageviewState extends State<Imageview> {
     final Color bgColor = themeChange.darkTheme ? Colors.black54 : Colors.white;
 
     return Scaffold(
-      body: _isLoading
-          ? const SpinKitRotatingCircle(
-              color: Colors.blue,
-            )
-          : SafeArea(
-              child: Column(
-                children: <Widget>[
-                  Expanded(
-                    flex: 10,
-                    child: Padding(
-                      padding: Space.all(0.75),
-                      child: Image.file(
-                        files[index],
+      body:
+          _isLoading
+              ? const SpinKitRotatingCircle(color: Colors.blue)
+              : SafeArea(
+                child: Column(
+                  children: <Widget>[
+                    Expanded(
+                      flex: 10,
+                      child: Padding(
+                        padding: Space.all(0.75),
+                        child: Image.file(files[index]),
                       ),
                     ),
-                  ),
-                  Expanded(
-                    child: Container(
-                      // padding: Space.all(),
-                      color: themeChange.darkTheme
-                          ? Colors.black87
-                          : Colors.blue[600],
-                      child: Row(
-                        mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                        children: <Widget>[
-                          TextButton(
-                            onPressed: () {
-                              _showChoiceDialogHome(context);
-                            },
-                            child: Column(
-                              mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                              children: <Widget>[
-                                Icon(
-                                  Icons.arrow_back,
-                                  color: Colors.white,
-                                  size: AppDimensions.font(
-                                    10,
-                                  ),
-                                ),
-                                Text(
-                                  S.back,
-                                  style: AppText.l1!.cl(Colors.white),
-                                ),
-                              ],
-                            ),
-                          ),
-                          Opacity(
-                            opacity: index == 0 ? 0.5 : 1,
-                            child: TextButton(
+                    Expanded(
+                      child: Container(
+                        color:
+                            themeChange.darkTheme
+                                ? Colors.black87
+                                : Colors.blue[600],
+                        child: Row(
+                          mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+                          children: <Widget>[
+                            TextButton(
                               onPressed: () {
-                                //Navigator.of(context).pop();
-                                if (index == 0) {
-                                  print("no undo possible");
-                                  //implement disabled undo if no undo is possible
-                                } else {
-                                  setState(() {
-                                    index--;
-                                    files.removeLast();
-                                    print(widget.list.imagelist.length);
-                                    // widget.list.imagepath.removeLast();
-                                  });
-                                }
+                                _showChoiceDialogHome(context);
                               },
                               child: Column(
                                 mainAxisAlignment:
                                     MainAxisAlignment.spaceEvenly,
                                 children: <Widget>[
                                   Icon(
-                                    Icons.undo,
+                                    Icons.arrow_back,
                                     color: Colors.white,
-                                    size: AppDimensions.font(
-                                      10,
-                                    ),
+                                    size: AppDimensions.font(10),
                                   ),
                                   Text(
                                     S.back,
@@ -289,91 +199,118 @@ class _ImageviewState extends State<Imageview> {
                                 ],
                               ),
                             ),
-                          ),
-                          TextButton(
-                            onPressed: () {
-                              if (files.isNotEmpty) {
-                                cropimage(files[index], appBarColor, bgColor);
-                              } else {
-                                cropimage(widget.file, appBarColor, bgColor);
-                              }
-                            },
-                            child: Column(
-                              mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                              children: <Widget>[
-                                Icon(
-                                  Icons.crop_rotate,
-                                  color: Colors.white,
-                                  size: AppDimensions.font(
-                                    10,
-                                  ),
+                            Opacity(
+                              opacity: index == 0 ? 0.5 : 1,
+                              child: TextButton(
+                                onPressed: () {
+                                  if (index == 0) {
+                                    print("no undo possible");
+                                  } else {
+                                    setState(() {
+                                      index--;
+                                      files.removeLast();
+                                      print(widget.list.imagelist.length);
+                                    });
+                                  }
+                                },
+                                child: Column(
+                                  mainAxisAlignment:
+                                      MainAxisAlignment.spaceEvenly,
+                                  children: <Widget>[
+                                    Icon(
+                                      Icons.undo,
+                                      color: Colors.white,
+                                      size: AppDimensions.font(10),
+                                    ),
+                                    Text(
+                                      S.back,
+                                      style: AppText.l1!.cl(Colors.white),
+                                    ),
+                                  ],
                                 ),
-                                Text(
-                                  S.crop,
-                                  style: AppText.l1!.cl(Colors.white),
-                                ),
-                              ],
+                              ),
                             ),
-                          ),
-                          TextButton(
-                            onPressed: () {
-                              getFilterImage(context, appBarColor);
-                            },
-                            child: Column(
-                              mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                              children: <Widget>[
-                                Icon(
-                                  Icons.filter,
-                                  color: Colors.white,
-                                  size: AppDimensions.font(
-                                    10,
+                            TextButton(
+                              onPressed: () {
+                                if (files.isNotEmpty) {
+                                  cropimage(files[index], appBarColor, bgColor);
+                                } else {
+                                  cropimage(widget.file, appBarColor, bgColor);
+                                }
+                              },
+                              child: Column(
+                                mainAxisAlignment:
+                                    MainAxisAlignment.spaceEvenly,
+                                children: <Widget>[
+                                  Icon(
+                                    Icons.crop_rotate,
+                                    color: Colors.white,
+                                    size: AppDimensions.font(10),
                                   ),
-                                ),
-                                Text(
-                                  S.filter,
-                                  style: AppText.l1!.cl(Colors.white),
-                                ),
-                              ],
-                            ),
-                          ),
-                          TextButton(
-                            onPressed: () {
-                              if (files.isNotEmpty) {
-                                widget.list.imagelist.add(files[index]);
-                                widget.list.imagepath.add(files[index].path);
-                              } else {
-                                widget.list.imagelist.add(widget.file);
-                                widget.list.imagepath.add(widget.file.path);
-                              }
-                              Navigator.of(context).pushNamed(
-                                RouteConstants.multiDelete,
-                                arguments: widget.list,
-                              );
-                            },
-                            child: Column(
-                              mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                              children: <Widget>[
-                                Icon(
-                                  Icons.arrow_forward,
-                                  color: Colors.white,
-                                  size: AppDimensions.font(
-                                    10,
+                                  Text(
+                                    S.crop,
+                                    style: AppText.l1!.cl(Colors.white),
                                   ),
-                                ),
-                                Text(
-                                  S.next,
-                                  style: AppText.l1!.cl(Colors.white),
-                                ),
-                              ],
+                                ],
+                              ),
                             ),
-                          ),
-                        ],
+                            TextButton(
+                              onPressed: () {
+                                getFilterImage(context, appBarColor);
+                              },
+                              child: Column(
+                                mainAxisAlignment:
+                                    MainAxisAlignment.spaceEvenly,
+                                children: <Widget>[
+                                  Icon(
+                                    Icons.filter,
+                                    color: Colors.white,
+                                    size: AppDimensions.font(10),
+                                  ),
+                                  Text(
+                                    S.filter,
+                                    style: AppText.l1!.cl(Colors.white),
+                                  ),
+                                ],
+                              ),
+                            ),
+                            TextButton(
+                              onPressed: () {
+                                if (files.isNotEmpty) {
+                                  widget.list.imagelist.add(files[index]);
+                                  widget.list.imagepath.add(files[index].path);
+                                } else {
+                                  widget.list.imagelist.add(widget.file);
+                                  widget.list.imagepath.add(widget.file.path);
+                                }
+                                Navigator.of(context).pushNamed(
+                                  RouteConstants.multiDelete,
+                                  arguments: widget.list,
+                                );
+                              },
+                              child: Column(
+                                mainAxisAlignment:
+                                    MainAxisAlignment.spaceEvenly,
+                                children: <Widget>[
+                                  Icon(
+                                    Icons.arrow_forward,
+                                    color: Colors.white,
+                                    size: AppDimensions.font(10),
+                                  ),
+                                  Text(
+                                    S.next,
+                                    style: AppText.l1!.cl(Colors.white),
+                                  ),
+                                ],
+                              ),
+                            ),
+                          ],
+                        ),
                       ),
                     ),
-                  ),
-                ],
+                  ],
+                ),
               ),
-            ),
     );
   }
 }
