+++ lib/utils/share_util.dart
@@ -1,3 +1,4 @@
+// ignore_for_file: deprecated_member_use
 import 'package:share_plus/share_plus.dart';
 import 'dart:io';
