+++ migration_plan.md
@@ -4,7 +4,7 @@
 Substituir todos os usos diretos da classe `Share` do pacote `share_plus` pelo novo utilitário `ShareUtil` para centralizar a funcionalidade de compartilhamento e lidar corretamente com os métodos obsoletos.
 
 ## Alterações Já Realizadas
-1. ✅ Criado o arquivo `share_util_new.dart` com métodos para compartilhar PDFs e texto
+1. ✅ Criado o arquivo `share_util.dart` com métodos para compartilhar PDFs e texto
 2. ✅ Atualizado `home.dart` para usar `ShareUtil.sharePdf` em vez de `Share.shareFiles`
 3. ✅ Atualizado `main_drawer.dart` para usar `ShareUtil.shareText` em vez de `Share.share`
 4. ✅ Atualizado `starred_documents.dart` para usar `ShareUtil.sharePdf` em vez de `Share.shareXFiles`
@@ -16,8 +16,7 @@ Substituir todos os usos diretos da classe `Share` do pacote `share_plus` pelo n
 1. ⬜ `image_view.dart`: O arquivo tem um método de menu popup comentado que usa a funcionalidade de compartilhamento. Se este código for descomentado no futuro, será necessário atualizá-lo para usar `ShareUtil.sharePdf()`.
 2. ⬜ Revisar estrutura do arquivo `image_view.dart` que parece ter alguns problemas de formatação.
 
-## Observações
-- O arquivo `share_util_new.dart` foi criado com a anotação `// ignore_for_file: deprecated_member_use` para evitar avisos de depreciação
+- O arquivo `share_util.dart` contém a anotação `// ignore_for_file: deprecated_member_use` para evitar avisos de depreciação
 - Todos os métodos incluem verificações de existência de arquivo antes de compartilhar para evitar erros
 - A API antiga (`Share.share`, `Share.shareFiles`, etc.) foi totalmente substituída pela nova API (`SharePlus.instance.share()` com `ShareParams`)
