+++ pubspec.yaml
@@ -11,44 +11,44 @@ dependencies:
   flutter:
     sdk: flutter
   animated_splash_screen: ^1.3.0
-  camera: ^0.10.5+5
-  cupertino_icons: ^1.0.5
+  camera: ^0.11.1
+  cupertino_icons: ^1.0.8
   double_back_to_close_app: ^2.1.0
-  easy_folder_picker: ^1.3.0
-  flutter_launcher_icons: ^0.13.1
-  flutter_spinkit: ^5.1.0
-  flutter_svg: ^2.0.9  # Versão atualizada
-  fluttertoast: ^8.2.1
+  easy_folder_picker: ^1.3.4
+  flutter_launcher_icons: ^0.14.4
+  flutter_spinkit: ^5.2.1
+  flutter_svg: ^2.2.0
+  fluttertoast: ^8.2.12
   image_gallery_saver: ^2.0.3  # Substituto moderno para gallery_saver
   hive: ^2.2.3
   hive_flutter: ^1.1.0
   image_cropper: ^9.1.0
   image_picker: ^1.1.2
   intl: 0.19.0  # Versão compatível com wiredash
-  introduction_screen: ^3.1.1
-  open_file: ^3.2.1
-  package_info_plus: ^5.0.1
-  path: ^1.8.2
-  path_provider: ^2.1.2
-  pdf: ^3.10.4
-  permission_handler: ^11.4.0  # Versão compatível com easy_folder_picker
+  introduction_screen: ^3.1.17
+  open_file: ^3.5.10
+  package_info_plus: ^8.3.0
+  path: ^1.9.1
+  path_provider: ^2.1.5
+  pdf: ^3.11.3
+  permission_handler: ^12.0.0+1
   image_editor_plus: ^1.0.6  # Substituto moderno para photofilters
-  printing: ^5.9.3
-  provider: ^6.1.2
-  quick_actions: ^1.0.1
+  printing: ^5.14.2
+  provider: ^6.1.5
+  quick_actions: ^1.1.0
   rating_dialog: ^2.0.4
-  share_plus: ^7.2.2
-  shared_preferences: ^2.2.2
-  source_span: ^1.10.0
-  url_launcher: ^6.2.5
-  wiredash: ^2.4.0
+  share_plus: ^11.0.0
+  shared_preferences: ^2.5.3
+  source_span: ^1.10.1
+  url_launcher: ^6.3.1
+  wiredash: ^2.5.0
 
 dev_dependencies:
-  build_runner: ^2.4.8
+  build_runner: ^2.5.0
   flutter_test:
     sdk: flutter
-  hive_generator: ^2.0.0
-  lint: ^2.2.0
+  hive_generator: ^2.0.1
+  lint: ^2.8.0
 
 flutter_icons:
   android: true
